Task 4.04 finished for OO

+++ src/main/java/br/ufpb/dce/aps/coffeemachine/impl/Cafe.java
@@ -2,23 +2,31 @@ package br.ufpb.dce.aps.coffeemachine.impl;
 
 import br.ufpb.dce.aps.coffeemachine.ComponentsFactory;
 import br.ufpb.dce.aps.coffeemachine.Messages;
+import br.ufpb.dce.aps.coffeemachine.Recipe;
 
-public class Cafe {
+public class Cafe{
 	
 	protected ComponentsFactory factory;
-	protected int agua;
-	protected int po;
-	protected int poRelease;
-	protected int precoCafe;
+	protected double agua;
+	protected double po;
+	protected double poRelease;
+	protected double precoCafe;
+//	protected Recipe receita;
 	
 
 	public Cafe(ComponentsFactory factory){
 		this.factory = factory;
-		agua = 100;
-		po = 15;
-		poRelease = 100;
-		precoCafe = 35;
+//		receita = new Recipe();
+		agua = 100.0;
+		po = 15.0;
+		poRelease = 100.0;
+		precoCafe = 35.0;
 	}
+	
+//	protected void setReceita(Recipe recipe){
+//		//agua = recipe.getIngredientQuantity("Water");
+//		po = recipe.getIngredientQuantity("Coffee Powder");
+//	}
 	
 	protected void servirBebida() {
 		factory.getDisplay().info(Messages.RELEASING);
@@ -65,7 +73,7 @@ public class Cafe {
 	}
 
 	protected void liberarPo(){
-		factory.getCoffeePowderDispenser().release(15);
+		factory.getCoffeePowderDispenser().release(15.0);
 	}
 	
 	protected void liberarCopo(){
@@ -73,10 +81,10 @@ public class Cafe {
 	}
 	
 	protected void liberarBebida(){
-		factory.getDrinkDispenser().release(100);
+		factory.getDrinkDispenser().release(100.0);
 	}
 
-	public int getPreco() {
+	public double getPreco() {
 		return precoCafe;
 	}
 }
